### 2025-11-03 #9343 @editorialbot
[REVIEW]: OnlinePCA.jl: A Julia Package for Out-of-core and Sparse Principal Component Analysis
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/OnlinePCA.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.10
