### 2025-12-02 0a7bb1f
Fix for JOSS review (docs and tests)
