+++ test/test_gd.jl
@@ -24,6 +24,7 @@ out_gd4 = gd(input=joinpath(tmp, "Data.zst"),
 	rowmeanlist=joinpath(dense_path, "Feature_FTTMeans.csv"),
 	logdir=dense_path)
 
+# Size tests
 @test size(out_gd1[1]) == (99, 3)
 @test size(out_gd1[2]) == (3, )
 @test size(out_gd1[3]) == (300, 3)
@@ -51,6 +52,41 @@ out_gd4 = gd(input=joinpath(tmp, "Data.zst"),
 @test size(out_gd4[4]) == (99, 3)
 @test size(out_gd4[5]) == ()
 @test size(out_gd4[6]) == ()
+
+# Accuracy tests:
+# eigenvalues should be non-negative and sorted in descending order
+## Non-negative eigenvalues
+@test all(out_gd1[2] .>= 0)
+@test all(out_gd2[2] .>= 0)
+@test all(out_gd3[2] .>= 0)
+@test all(out_gd4[2] .>= 0)
+
+## Descending order (PC1 > PC2 > PC3)
+@test issorted(out_gd1[2], rev=true)
+@test issorted(out_gd2[2], rev=true)
+@test issorted(out_gd3[2], rev=true)
+@test issorted(out_gd4[2], rev=true)
+
+## Loadings should have unit norm (columns are orthonormal)
+for j in 1:3
+    @test isapprox(norm(out_gd1[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_gd2[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_gd3[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_gd4[3][:, j]), 1.0, atol=0.1)
+end
+
+## Total variance explained should be positive
+@test sum(out_gd1[2]) > 0
+@test sum(out_gd2[2]) > 0
+@test sum(out_gd3[2]) > 0
+@test sum(out_gd4[2]) > 0
+#####################################
 #####################################
 
 #####################################
