+++ test/test_singlepass.jl
@@ -5,11 +5,28 @@ out_singlepass = singlepass(input=joinpath(tmp, "Data.zst"),
 	rowmeanlist=joinpath(dense_path, "Feature_FTTMeans.csv"),
 	logdir=dense_path)
 
+# Size tests
 @test size(out_singlepass[1]) == (99, 3)
 @test size(out_singlepass[2]) == (3, )
 @test size(out_singlepass[3]) == (300, 3)
 @test size(out_singlepass[4]) == (99, 3)
 @test size(out_singlepass[5]) == ()
+
+# Accuracy tests:
+# eigenvalues should be non-negative and sorted in descending order
+## Non-negative eigenvalues
+@test all(out_singlepass[2] .>= 0)
+
+## Descending order (PC1 > PC2 > PC3)
+@test issorted(out_singlepass[2], rev=true)
+
+## Loadings should have unit norm (columns are orthonormal)
+for j in 1:3
+    @test isapprox(norm(out_singlepass[3][:, j]), 1.0, atol=0.1)
+end
+
+## Total variance explained should be positive
+@test sum(out_singlepass[2]) > 0
 ####################################
 
 #####################################
