+++ test/test_svrg.jl
@@ -24,6 +24,7 @@ out_svrg4 = svrg(input=joinpath(tmp, "Data.zst"),
 	rowmeanlist=joinpath(dense_path, "Feature_FTTMeans.csv"),
 	logdir=dense_path)
 
+# Size tests
 @test size(out_svrg1[1]) == (99, 3)
 @test size(out_svrg1[2]) == (3, )
 @test size(out_svrg1[3]) == (300, 3)
@@ -51,6 +52,40 @@ out_svrg4 = svrg(input=joinpath(tmp, "Data.zst"),
 @test size(out_svrg4[4]) == (99, 3)
 @test size(out_svrg4[5]) == ()
 @test size(out_svrg4[6]) == ()
+
+# Accuracy tests:
+# eigenvalues should be non-negative and sorted in descending order
+## Non-negative eigenvalues
+@test all(out_svrg1[2] .>= 0)
+@test all(out_svrg2[2] .>= 0)
+@test all(out_svrg3[2] .>= 0)
+@test all(out_svrg4[2] .>= 0)
+
+## Descending order (PC1 > PC2 > PC3)
+@test issorted(out_svrg1[2], rev=true)
+@test issorted(out_svrg2[2], rev=true)
+@test issorted(out_svrg3[2], rev=true)
+@test issorted(out_svrg4[2], rev=true)
+
+## Loadings should have unit norm (columns are orthonormal)
+for j in 1:3
+    @test isapprox(norm(out_svrg1[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_svrg2[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_svrg3[3][:, j]), 1.0, atol=0.1)
+end
+for j in 1:3
+    @test isapprox(norm(out_svrg4[3][:, j]), 1.0, atol=0.1)
+end
+
+## Total variance explained should be positive
+@test sum(out_svrg1[2]) > 0
+@test sum(out_svrg2[2]) > 0
+@test sum(out_svrg3[2]) > 0
+@test sum(out_svrg4[2]) > 0
 #####################################
 
 #####################################
