### 2025-12-02 8f61d4e
Fix for JOSS review (docs and tests)
