+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"commandline/#OnlinePCA.jl-(Command-line-tool)","page":"Command line Tool","title":"OnlinePCA.jl (Command line tool)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"All functions can be performed as command line tool in shell window and same options in OnlinePCA.jl (Julia API) are available.","category":"page"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"After installation of OnlinePCA.jl, command line tools are saved at YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/.","category":"page"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"The functions can be performed as below.","category":"page"},{"location":"commandline/#Binarization-(CSV-file)","page":"Command line Tool","title":"Binarization (CSV file)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"shell> julia YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/csv2bin \\\n--csvfile Data.csv \\\n--binfile OUTDIR/Data.zst","category":"page"},{"location":"commandline/#Binarization-(Matrix-Market-MM-file)","page":"Command line Tool","title":"Binarization (Matrix Market <MM> file)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"shell> julia YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/mm2bin \\\n--mmfile Data.mtx \\\n--binfile OUTDIR/Data.mtx.zst","category":"page"},{"location":"commandline/#Binarization-(Binary-COO-BinCOO-file)","page":"Command line Tool","title":"Binarization (Binary COO <BinCOO> file)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"shell> julia YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/bincoo2bin \\\n--mmfile Data.bincoo \\\n--binfile OUTDIR/Data.bincoo.zst","category":"page"},{"location":"commandline/#Summarization-(CSV-file)","page":"Command line Tool","title":"Summarization (CSV file)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"shell> julia YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/sumr \\\n--binfile OUTDIR/Data.zst \\\n--outdir OUTDIR \\\n--pseudocount 1f0","category":"page"},{"location":"commandline/#Summarization-(MM-file)","page":"Command line Tool","title":"Summarization (MM file)","text":"","category":"section"},{"location":"commandline/","page":"Command line Tool","title":"Command line Tool","text":"shell> julia YOUR_HOME_DIR/.julia/v1.x/OnlinePCA/bin/sumr \\\n--binfile OUTDIR/Data.zst \\\n--outdir OUTDIR \\\n--pseudocount 1f0 \\\n--sparse
[…]
