### 2025-12-02 c131995
Fix for JOSS review (docs and tests)
