### 2025-12-02 b871559
Fix for JOSS review (docs and tests)
