+++ paper/paper.md
@@ -15,30 +15,26 @@ affiliations:
    index: 1
  - name: Laboratory for Bioinformatics Research, RIKEN Center for Biosystems Dynamics Research, Japan
    index: 2
-date: 1 July 2025
+date: 2 December 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-Principal Component Analysis (PCA) is a widely used dimensionality reduction technique for identifying a small number of components that capture the maximum variance in high-dimensional data [@review1; @review2]. PCA has been applied across various fields of data science, including face recognition [@eigenface], animal behavior analysis [@eigenworm], omics studies [@review1; @review2; @onlinepcajl], population genetics [@genetics; @pcaone], and molecular dynamics simulations [@md].
-
-Despite its broad applicability, PCA becomes computationally prohibitive for large data matrices, making it difficult to apply in practice. In particular, recent advances in single-cell omics have led to datasets with millions of cells, for which standard PCA implementations often fail to scale. To meet this requirement, I originally developed \texttt{OnlinePCA.jl}, which is a Julia package to perform some PCA algorithms (\url{https://github.com/rikenbit/OnlinePCA.jl}).
+Principal Component Analysis (PCA) is a widely used dimensionality reduction technique [@review1; @review2], but becomes computationally prohibitive for large data matrices. Recent advances in single-cell omics have led to datasets with millions of cells, for which standard PCA implementations often fail to scale. \texttt{OnlinePCA.jl} is a Julia package that addresses this challenge by providing scalable PCA algorithms (\url{https://github.com/rikenbit/OnlinePCA.jl}).
 
 # Statement of need
 
-PCA is a workhorse algorithm for most data science tasks. However, as the size of the data matrix increases, it often becomes too large to fit into memory. In such cases, an out-of-core (OOC) implementation — where only subsets of data stored on disk are loaded into memory for computation — is desirable. Additionally, representing the data in a sparse matrix format, where only non-zero values and their coordinates are stored, is computationally advantageous [@onlinepcajl]. Therefore, a PCA implementation that supports both OOC computation and sparse data handling is highly desirable.
+PCA is widely used across diverse fields including face recognition [@eigenface], animal behavior [@eigenworm], genomics [@review1; @review2], population genetics [@genetics; @pcaone], and molecular dynamics [@md]. However, large data matrices often exceed available memory. An out-of-core (OOC) implementation—where data subsets are loaded from disk—combined with sparse matrix support is highly desirable [@onlinepcajl].
 
 ## New features since version 0.3.0
 
-\texttt{OnlinePCA.jl} has been provided some OOC PCA functions such as \texttt{oja}, \texttt{ccipca}, \texttt{gd}, \texttt{rsgd}, \texttt{svrg}, \texttt{rsvrg}, \texttt{orthiter}, \texttt{arnoldi}, \texttt{lanczos}, \texttt{halko}, \texttt{algorithm971}, \texttt{rbkiter}, \texttt{singlepass}, and \texttt{singlepass2} and an OOC/Sparse PCA function (\texttt{tenxpca}) only for a sparse matrix in HDF5 files generated by 10X Cell Ranger (10X-HDF5) [@onlinepcajl]. These implementations were designed to take "CSV" files containing "short and fat" matrices—that is, matrices with relatively few rows (samples) and a large number of columns (features)—as typically seen in large-scale single-cell datasets.
[…]
