### 2025-12-02 #9343 @shravanngoswamii
@kokitsuyuzaki Is there any reason for calling [Binarization function](https://github.com/rikenbit/OnlinePCA.jl/blob/b871559b12acecc1fe74710714cf23276d3255c7/paper/paper.md?plain=1#L77C1-L86C39) `csv2bin` twice in the section `Preprocess of CSV`?
