### 2025-12-02 30ba09a
Fix for JOSS review (docs and tests)
