### 2025-12-02 #9343 @kokitsuyuzaki
@shravanngoswamii Thank you for the prompt response.
I have addressed the issues in v0.3.12.
https://github.com/rikenbit/OnlinePCA.jl/compare/v0.3.11...v0.3.12

> csv2bin twice

Sorry, this was just a typo. I also found the same typo in README.md and fixed it.

> scale="ftt" and rowmeanlist is Feature_LogMeans.csv

This was also a typo. Since the results were based on log transformation, I added scale="log" as well.
