### 2025-12-03 #9343 @shravanngoswamii
One more thing, CI is only tested against Julia version `1.10`, and in many Julia packages, CI was failing for Julia version `1.12` -- This is not part of review for JOSS, but still, must have a look.
