### 2025-12-03 e0b1538
Fix for JOSS review (docs and tests)
