+++ .github/workflows/CI.yml
@@ -24,6 +24,7 @@ jobs:
       matrix:
         version:
           - '1.10'
+          - '1'
         os:
           - ubuntu-latest
         arch:
