### 2025-12-03 5f477a1
Fix for JOSS review (docs and tests)
