### 2025-12-09 9b47055
Fix for JOSS review (docs and tests)
