### 2026-01-23 #9343 @crvernon
:wave: Hello all.  It looks like both @sa-nouri and @shravanngoswamii have completed their checklists.  @sa-nouri could you respond to the [author's comment ](https://github.com/openjournals/joss-reviews/issues/9343#issuecomment-3765056464)and let me know if you are OK with this submission moving forward?  Thank you.
