### 2026-01-23 #9343 @crvernon
👋  @kokitsuyuzaki  - I need you to make the following in your paper:

- [x] LINE 64:  "...see the Figure 2 in (Tsuyuzaki, 2020)." should be cited in-text as "...see the Figure 2 in Tsuyuzaki (2020)."
- [x] LINE 74: The "d" in "dask" should be capitalized.  You can do this in your bib file by wrapping any text you want to maintain the formatting of in curly brackets.
- [x] LINE 77:  The "e" in "europe" should be capitalized.
- [x] LINE 82:  The "p" in "python" should be capitalized.

- [x]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.

- [x]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).

- [x]  Please respond with the DOI of the archived version here along with the version tag name released.
