+++ paper/paper.bib
@@ -30,7 +30,7 @@
 @article{genetics,
   author = {Novembre, J. et al.},
   journal = {Nature},
-  title = {Genes mirror geography within Europe},
+  title = {Genes mirror geography within {E}urope},
   volume = {456},
   pages = {98-101},
   year = {2008},
@@ -80,7 +80,7 @@
 @article{sklearn,
   author = {Pedregosa, F. et al.},
   journal = {Journal of Machine Learning Research},
-  title = {Scikit-learn: Machine learning in Python},
+  title = {Scikit-learn: Machine learning in {P}ython},
   volume = {12(85)},
   pages = {2825–2830},
   year = {2011}}
@@ -89,7 +89,7 @@
 @article{dask,
   author = {Moreno, M. et al.},
   journal = {BMC Bioinformatics},
-  title = {Scalable transcriptomics analysis with Dask: applications in data science and machine learning},
+  title = {Scalable transcriptomics analysis with {D}ask: applications in data science and machine learning},
   volume = {23(514)},
   year = {2022},
   doi = {10.1186/s12859-022-05065-3}}
