+++ Project.toml
@@ -2,7 +2,7 @@ name = "OnlinePCA"
 uuid = "975b03d1-ab45-5730-8389-540698ac00dc"
 license = "MIT"
 authors = ["Koki Tsuyuzaki <koki.tsuyuzaki@gmail.com>"]
-version = "0.3.13"
+version = "0.3.14"
 
 [deps]
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
