### 2026-01-29 #9343 @kokitsuyuzaki
@crvernon Sorry for the delay, and thank you for your guidance.

I have fully addressed all of the comments.

- I’ve created a new release: [v0.3.14](https://github.com/rikenbit/OnlinePCA.jl/releases/tag/v0.3.14)
- The Zenodo archive is available here: https://zenodo.org/records/18415366
  DOI: [10.5281/zenodo.18415366](https://doi.org/10.5281/zenodo.18415366)

The paper title and author list in the Zenodo metadata match the JOSS paper exactly.

Please let me know if anything else is needed!
