### 2026-01-29 #9343 @crvernon
🥳 Congratulations on your new publication @kokitsuyuzaki ! Many thanks to @sa-nouri and @shravanngoswamii for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
