### 2026-01-30 #9343 @kokitsuyuzaki
Thank you so much @crvernon for your kind support throughout the process, and many thanks to @sa-nouri and @shravanngoswamii for your thoughtful and constructive reviews! I truly appreciate your time and expertise.

I'd be happy to consider serving as a reviewer for JOSS — it's a great community and review process!
